Split article into fragments
